revert: 回滚 partial 设计
