+++ src/components/BootstrapBlazor.DockView/Components/DockViewComponent.razor
@@ -1,0 +1,18 @@
+@inherits DockViewComponentBase
+
+<div id="@Id" class="bb-dockview-panel" data-bb-key="@Key" data-bb-title="@Title">
+    @if (TitleTemplate != null)
+    {
+        <div class="bb-dockview-item-title">
+            @TitleTemplate
+        </div>
+    }
+    else if (ShowTitleBar)
+    {
+        <DockViewTitleBar BarIcon="@TitleBarIcon" BarIconUrl="@TitleBarIconUrl" OnClickBarCallback="OnClickBar"></DockViewTitleBar>
+    }
+    @if (Visible)
+    {
+        @ChildContent
+    }
+</div>
